+++ src/routes/users.routes.js
@@ -1,0 +1,5 @@
+import { Router } from "express";
+
+const router = Router();
+
+export default router;
